Rename update estimators to estimators
