+++ reinforcelab/agents/agent.py
@@ -5,7 +5,7 @@ from typing import Union
 from abc import ABC, abstractmethod
 
 from reinforcelab.brains import Brain
-from reinforcelab.update_estimators import UpdateEstimator
+from reinforcelab.estimators import Estimator
 from reinforcelab.memory_buffers import MemoryBuffer
 from reinforcelab.action_selectors import ActionSelector
 from reinforcelab.experience import Experience
