+++ reinforcelab/agents/value_optimization/tabular/sarsa.py
@@ -2,7 +2,7 @@ from gymnasium import Env
 
 from reinforcelab.agents.agent import Agent
 from reinforcelab.brains import QTable
-from reinforcelab.update_estimators import SARSAEstimator
+from reinforcelab.estimators import SARSAEstimator
 from reinforcelab.action_selectors import EpsilonGreedy
 from reinforcelab.memory_buffers import OrderedBuffer
 from reinforcelab.utils import get_state_action_sizes
