+++ reinforcelab/estimators/__init__.py
@@ -1,9 +1,9 @@
-from .update_estimator import UpdateEstimator
+from .estimator import Estimator
 from .max_q_estimator import MaxQEstimator
 from .double_q_estimator import DoubleQEstimator
 from .sarsa_estimator import SARSAEstimator
 from .sars_estimator import SARSEstimator
 from .expected_sarsa_estimator import ExpectedSARSAEstimator
 
-__all__ = [UpdateEstimator, MaxQEstimator, DoubleQEstimator,
+__all__ = [Estimator, MaxQEstimator, DoubleQEstimator,
            SARSAEstimator, SARSEstimator, ExpectedSARSAEstimator]
