+++ reinforcelab/estimators/max_q_estimator.py
@@ -3,33 +3,21 @@ import torch
 from torch import Tensor
 import gymnasium as gym
 
-from .update_estimator import UpdateEstimator
+from .estimator import Estimator
 from reinforcelab.experience import Experience
 from reinforcelab.brains import Brain
 from reinforcelab.utils import space_is_type
 
 
-class MaxQEstimator(UpdateEstimator):
+class MaxQEstimator(Estimator):
     def __init__(self, env: gym.Env, gamma: float, transforms=None):
         """Creates a Q estimator
 
         Args:
             gamma (float): Gamma parameter or discount factor
         """
-        self.__validate_env(env)
         self.gamma = gamma
         self.transforms = transforms
-
-    def __validate_env(self, env: gym.Env):
-        """Determines if the environment is compatible with the estimator.
-        If the action space is not Dsicrete, raises an error
-
-        Args:
-            env (gym.Env): Gym environment
-        """
-        act_disc = space_is_type(env.action_space, gym.spaces.Discrete)
-        if not act_disc:
-            raise RuntimeError("Incompatible action space")
 
     def __call__(self, experience: Experience, brain: Brain) -> Tensor:
         """Computes the action-value estimation for an experience tuple with the given local and
